interop::clients::corecrypto::web

Function find_js_functions

fn find_js_functions(source: &str) -> HashMap<String, String>

Parse the source, returning all function bodies by function name

This contains a bunch of potentially panicing code, because there’s no nice way to do a one-time init without such as things stand right now.

The most likely cause of a panic here is if the input file cannot be parsed.