Trait interop::mu_server::Mu

pub trait Mu: Send + Sync + 'static {
    type Error: CustomError + Send + Sync + 'static;
    type MuMuteValidationType: Send + Sync;
    type MuMuValidationType: Send + Sync;
    fn mu<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<Ping>
    ) -> Pin<Box<dyn Future<Output = Result<Pong, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn mu_mute_validation<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<Option<Ping>>
    ) -> Pin<Box<dyn Future<Output = Result<Self::MuMuteValidationType, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn mu_mute<'life0, 'async_trait>(
        &'life0 self, 
        validation_value: Self::MuMuteValidationType, 
        socket: Socket<Ping, Pong>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn mu_mu_validation<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<Option<Ping>>
    ) -> Pin<Box<dyn Future<Output = Result<Self::MuMuValidationType, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn mu_mu<'life0, 'async_trait>(
        &'life0 self, 
        validation_value: Self::MuMuValidationType, 
        socket: Socket<Ping, Pong>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn middleware(&self, endpoint: &'static str) -> BoxedFilter<()> { ... }

    fn mu_middleware(&self, endpoint: &'static str) -> BoxedFilter<()> { ... }

    fn mu_mute_middleware(&self, endpoint: &'static str) -> BoxedFilter<()> { ... }

    fn mu_mute_on_upgrade(&self, response: Response) -> Response { ... }

    fn mu_mu_middleware(&self, endpoint: &'static str) -> BoxedFilter<()> { ... }

    fn mu_mu_on_upgrade(&self, response: Response) -> Response { ... }
}

Generated trait containing hRPC methods that should be implemented for use with MuServer.

Associated Types

type Error: CustomError + Send + Sync + 'static

type MuMuteValidationType: Send + Sync

type MuMuValidationType: Send + Sync

Required methods

fn mu<'life0, 'async_trait>(
&'life0 self,
request: Request<Ping>
) -> Pin<Box<dyn Future<Output = Result<Pong, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Mu RPC.

fn mu_mute_validation<'life0, 'async_trait>(
&'life0 self,
request: Request<Option<Ping>>
) -> Pin<Box<dyn Future<Output = Result<Self::MuMuteValidationType, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

The message will always be None.

fn mu_mute<'life0, 'async_trait>(
&'life0 self,
validation_value: Self::MuMuteValidationType,
socket: Socket<Ping, Pong>
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

MuMute RPC.

fn mu_mu_validation<'life0, 'async_trait>(
&'life0 self,
request: Request<Option<Ping>>
) -> Pin<Box<dyn Future<Output = Result<Self::MuMuValidationType, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

The message may be None or Some.

fn mu_mu<'life0, 'async_trait>(
&'life0 self,
validation_value: Self::MuMuValidationType,
socket: Socket<Ping, Pong>
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

MuMu RPC. ``` test ```

Provided methods

fn middleware(&self, endpoint: &'static str) -> BoxedFilter<()>

Filter to be run before all API operations but after API path is matched, for all endpoints.

fn mu_middleware(&self, endpoint: &'static str) -> BoxedFilter<()>

Filter to be run before all API operations but after API path is matched.

fn mu_mute_middleware(&self, endpoint: &'static str) -> BoxedFilter<()>

Filter to be run before all API operations but after API path is matched.

fn mu_mute_on_upgrade(&self, response: Response) -> Response

Method that can be used to modify the response sent when the WebSocket is upgraded.

fn mu_mu_middleware(&self, endpoint: &'static str) -> BoxedFilter<()>

Filter to be run before all API operations but after API path is matched.

fn mu_mu_on_upgrade(&self, response: Response) -> Response

Method that can be used to modify the response sent when the WebSocket is upgraded.