Trait interop::mu_server::Mu

pub trait Mu: Clone + Send + Sized + 'static {
    fn mu(
        &mut self, 
        request: HrpcRequest<Ping>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<Pong>>>;

    fn mu_mute(
        &mut self, 
        request: HrpcRequest<()>, 
        socket: Socket<Ping, Pong>
    ) -> BoxFuture<'_, ServerResult<()>>;

    fn mu_mu(
        &mut self, 
        request: HrpcRequest<()>, 
        socket: Socket<Ping, Pong>
    ) -> BoxFuture<'_, ServerResult<()>>;

    fn mu_middleware(&self, endpoint: &'static str) -> Option<HrpcLayer> { ... }

    fn mu_mute_middleware(&self, endpoint: &'static str) -> Option<HrpcLayer> { ... }

    fn mu_mute_on_upgrade(&mut self, response: HttpResponse) -> HttpResponse { ... }

    fn mu_mu_middleware(&self, endpoint: &'static str) -> Option<HrpcLayer> { ... }

    fn mu_mu_on_upgrade(&mut self, response: HttpResponse) -> HttpResponse { ... }
}

Generated trait containing hRPC methods that should be implemented for use with MuServer.

Required methods

fn mu(
&mut self,
request: HrpcRequest<Ping>
) -> BoxFuture<'_, ServerResult<HrpcResponse<Pong>>>

Mu RPC.

fn mu_mute(
&mut self,
request: HrpcRequest<()>,
socket: Socket<Ping, Pong>
) -> BoxFuture<'_, ServerResult<()>>

MuMute RPC.

fn mu_mu(
&mut self,
request: HrpcRequest<()>,
socket: Socket<Ping, Pong>
) -> BoxFuture<'_, ServerResult<()>>

MuMu RPC. ``` test ```

Provided methods

fn mu_middleware(&self, endpoint: &'static str) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn mu_mute_middleware(&self, endpoint: &'static str) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn mu_mute_on_upgrade(&mut self, response: HttpResponse) -> HttpResponse

Method that can be used to modify the response sent when the WebSocket is upgraded.

fn mu_mu_middleware(&self, endpoint: &'static str) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn mu_mu_on_upgrade(&mut self, response: HttpResponse) -> HttpResponse

Method that can be used to modify the response sent when the WebSocket is upgraded.