Trait interop::mu_server::Mu

pub trait Mu: Send + Sync + 'static {
    fn mu<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<Ping>
    ) -> Pin<Box<dyn Future<Output = Result<Response<Pong>, ServerError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn mu_mute<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<()>, 
        socket: Socket<Ping, Pong>
    ) -> Pin<Box<dyn Future<Output = Result<(), ServerError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn mu_mu<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<()>, 
        socket: Socket<Ping, Pong>
    ) -> Pin<Box<dyn Future<Output = Result<(), ServerError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn middleware(&self) -> HrpcRouteLayer { ... }

    fn mu_middleware(&self, endpoint: &'static str) -> HrpcLayer { ... }

    fn mu_mute_middleware(&self, endpoint: &'static str) -> HrpcLayer { ... }

    fn mu_mute_on_upgrade(
        &self, 
        response: Response<BoxBody>
    ) -> Response<BoxBody> { ... }

    fn mu_mu_middleware(&self, endpoint: &'static str) -> HrpcLayer { ... }

    fn mu_mu_on_upgrade(&self, response: Response<BoxBody>) -> Response<BoxBody> { ... }
}

Generated trait containing hRPC methods that should be implemented for use with MuServer.

Required methods

fn mu<'life0, 'async_trait>(
&'life0 self,
request: Request<Ping>
) -> Pin<Box<dyn Future<Output = Result<Response<Pong>, ServerError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Mu RPC.

fn mu_mute<'life0, 'async_trait>(
&'life0 self,
request: Request<()>,
socket: Socket<Ping, Pong>
) -> Pin<Box<dyn Future<Output = Result<(), ServerError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

MuMute RPC.

fn mu_mu<'life0, 'async_trait>(
&'life0 self,
request: Request<()>,
socket: Socket<Ping, Pong>
) -> Pin<Box<dyn Future<Output = Result<(), ServerError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

MuMu RPC. ``` test ```

Provided methods

fn middleware(&self) -> HrpcRouteLayer

Filter to be run before all API operations but after API path is matched, for all endpoints.

fn mu_middleware(&self, endpoint: &'static str) -> HrpcLayer

Filter to be run before all API operations but after API path is matched.

fn mu_mute_middleware(&self, endpoint: &'static str) -> HrpcLayer

Filter to be run before all API operations but after API path is matched.

fn mu_mute_on_upgrade(&self, response: Response<BoxBody>) -> Response<BoxBody>

Method that can be used to modify the response sent when the WebSocket is upgraded.

fn mu_mu_middleware(&self, endpoint: &'static str) -> HrpcLayer

Filter to be run before all API operations but after API path is matched.

fn mu_mu_on_upgrade(&self, response: Response<BoxBody>) -> Response<BoxBody>

Method that can be used to modify the response sent when the WebSocket is upgraded.